+++ src/App/App.tsx
@@ -3,15 +3,19 @@ import useJSS from './style'
 import { RightDrawer, LeftDrawer, ModuleViewFill } from '../components/all'
 import { useSelector } from 'react-redux'
 import { RootState, ContainerModule } from '../redux/stateTSTypes'
+import { AudioModules } from '../audioModules/moduleTypes'
 
 declare global {
   interface Window { 
     fillContainerID: string
+    audioModules: AudioModules
   }
 }
 
 window.highlightedID = 'project' // delete this, for dev
 window.fillContainerID = 'project'
+
+window.audioModules = {}
 
 function App() {
   const classes = useJSS()
