--- src/modules/moduleTypes.ts
@@ -1,6 +1,0 @@
-export const GATE = 'GATE'
-export const OSCILLATOR = 'OSCILLATOR'
-export const CONTAINER = 'CONTAINER'
-export const OUTPUT = 'OUTPUT'
-
-export type ModuleType = 'GATE' | 'OSCILLATOR' | 'CONTAINER' | 'OUTPUT'
