+++ src/redux/modules/reducers/addModule.ts
@@ -1,6 +1,6 @@
 import { Modules, ContainerModule } from '../../stateTSTypes'
 import { AddModuleAction } from '../moduleTSTypes'
-import { CONTAINER, OSCILLATOR, GATE, OUTPUT } from '../../../modules/moduleTypes'
+import { CONTAINER, OSCILLATOR, GATE, OUTPUT } from '../../../audioModules/moduleTypes'
 
 const addModule = (state: Modules, { id, moduleType, row, col, parentID }: AddModuleAction) => {
   switch(moduleType) {
